+++ src/io/io.c
@@ -13,7 +13,7 @@ size_t readFile(const char *filename, u8 **buffer) {
 		return 0;
 	}
 	fseek(file, 0, SEEK_END);
-	const long size = ftell(file);
+	const size_t size = (size_t) ftell(file);
 	*buffer = malloc(size);
 	rewind(file);
 	fread(*buffer, 1, size, file);
@@ -37,7 +37,7 @@ NO_RETURN void fail(const char *RESTRICT format, ...) {
 }
 NO_RETURN void failWithFileError(const char *filename, const char *invocation) {
 	const char *format = "%s: Cannot access \"%s\"";
-	const size_t messageLength = snprintf(NULL, 0, format, invocation, filename);
+	const size_t messageLength = (size_t) snprintf(NULL, 0, format, invocation, filename);
 	char *message = malloc(messageLength);
 	snprintf(message, messageLength, format, invocation, filename);
 	perror(message);
